projects/django-vue/back/requirements.txt
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djoser==2.1.0
drf-yasg==1.20.0
PyYAML==6.0
loguru
psycopg2
requests==2.27.1
